requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
torch
torchvision
Pillow>=9.0.0
